The functions are separated into different modules. Missing data types of function parameters are specified. The alphabetical order in imports.
